Create Find_digits
Given an integer, , traverse its digits (1,2,...,n) and determine how many digits evenly divide  (i.e.: count the number of times  divided by each digit i has a remainder of ). Print the number of evenly divisible digits.

Note: Each digit is considered to be unique, so each occurrence of the same evenly divisible digit should be counted (i.e.: for , the answer is ).

Input Format
The first line is an integer, , indicating the number of test cases. 
The  subsequent lines each contain an integer, .

Output Format
For every test case, count and print (on a new line) the number of digits in  that are able to evenly divide