Create Sherlock_and_Squares
Watson gives two integers ( and ) to Sherlock and asks if he can count the number of square integers between  and  (both inclusive).

Note: A square integer is an integer which is the square of any integer. For example, 1, 4, 9, and 16 are some of the square integers as they are squares of 1, 2, 3, and 4, respectively.

Input Format 
The first line contains , the number of test cases.  test cases follow, each in a new line. 
Each test case contains two space-separated integers denoting  and .

Output Format 
For each test case, print the required answer in a new line.